feat(make-test-sound): rename file
