+++ test/make-test-sound.d.ts
